[front/components/actions/mcp] - refactor: rename CreateMCPServerDialog to CreateMCPServerSheet
 - Renamed CreateMCPServerDialog component for consistency with other Sheet components
 - Updated import references in AdminActionsList to reflect the new component name
