+++ front/components/actions/mcp/AdminActionsList.tsx
@@ -9,7 +9,7 @@ import type { CellContext, ColumnDef } from "@tanstack/react-table";
 import { useMemo, useState } from "react";
 
 import { AddActionMenu } from "@app/components/actions/mcp/AddActionMenu";
-import { CreateMCPServerSheet } from "@app/components/actions/mcp/CreateMCPServerDialog";
+import { CreateMCPServerSheet } from "@app/components/actions/mcp/CreateMCPServerSheet";
 import { AssistantDetails } from "@app/components/assistant/details/AssistantDetails";
 import { ACTION_BUTTONS_CONTAINER_ID } from "@app/components/spaces/SpacePageHeaders";
 import { UsedByButton } from "@app/components/spaces/UsedByButton";
